move scenario tests into rds
